Inverse Kinematics Update
I put the equation you sent me into the invkin function.

I also set up a dynamic decision on what direction the wheels should spin.

On top of that I commented out the previous IK equations and I also put a version of the IK with decimal values and commented this out as well.